refactor: reorganized folders structure
