+++ src/app.go
@@ -1,13 +1,13 @@
 package main
 
 import (
-	"comiditapp/api/controllers/any_role/auth"
-	client "comiditapp/api/controllers/any_role/clients"
-	"comiditapp/api/controllers/any_role/home"
-	"comiditapp/api/controllers/any_role/profile"
-	"comiditapp/api/controllers/any_role/restaurants"
-	"comiditapp/api/controllers/restaurant_role/orders"
-	"comiditapp/api/controllers/restaurant_role/products"
+	auth "comiditapp/api/controllers/auth"
+	clients "comiditapp/api/controllers/clients"
+	home "comiditapp/api/controllers/home"
+	orders "comiditapp/api/controllers/orders"
+	products "comiditapp/api/controllers/products"
+	profiles "comiditapp/api/controllers/profile"
+	restaurants "comiditapp/api/controllers/restaurants"
 	"comiditapp/api/database"
 	"comiditapp/api/env"
 	"log"
@@ -27,14 +27,11 @@ func main() {
 	r := gin.Default()
 	r.Use(cors.Default())
 
-	// any role users
 	home.Routes(r)
 	auth.Routes(r, db)
 	restaurants.Routes(r, db)
-	client.Routes(r, db)
-	profile.Routes(r, db)
-
-	// only restaurant role users
+	clients.Routes(r, db)
+	profiles.Routes(r, db)
 	orders.Routes(r, db)
 	products.Routes(r, db)
