+++ src/controllers/home/home.go
@@ -1,0 +1,15 @@
+package controllers
+
+import (
+	handlers "comiditapp/api/handlers/home"
+
+	"github.com/gin-gonic/gin"
+)
+
+func Routes(r *gin.Engine) {
+	homeGroup := r.Group("/")
+	{
+		homeGroup.GET("", handlers.GetHome)
+		homeGroup.GET("/home", handlers.GetHome)
+	}
+}
