--- src/handlers/any_role/auth/post_signout.go
@@ -1,14 +1,0 @@
-package auth_handler
-
-import (
-	repository "comiditapp/api/database/repositories/users"
-
-	"github.com/gin-gonic/gin"
-)
-
-func SignOutUser(repository *repository.MongoUsersRepository) gin.HandlerFunc {
-	return func(context *gin.Context) {
-		statusCode, response := repository.SignOutUser(context)
-		context.IndentedJSON(statusCode, response)
-	}
-}
