--- src/handlers/any_role/clients/get_client_by_id.go
@@ -1,14 +1,0 @@
-package client_handler
-
-import (
-	repository "comiditapp/api/database/repositories/users"
-
-	"github.com/gin-gonic/gin"
-)
-
-func GetClientById(repository *repository.MongoUsersRepository) gin.HandlerFunc {
-	return func(context *gin.Context) {
-		statusCode, response := repository.GetClientById(context)
-		context.IndentedJSON(statusCode, response)
-	}
-}
