--- src/handlers/any_role/home/get_home.go
@@ -1,11 +1,0 @@
-package home_handler
-
-import (
-	"net/http"
-
-	"github.com/gin-gonic/gin"
-)
-
-func GetHome(c *gin.Context) {
-	c.String(http.StatusOK, "Welcome to comiditapp homepage!! 🍔")
-}
