--- src/handlers/any_role/profile/get_order_by_id.go
@@ -1,14 +1,0 @@
-package profile_handler
-
-import (
-	repository "comiditapp/api/database/repositories/orders"
-
-	"github.com/gin-gonic/gin"
-)
-
-func GetOrderById(repository *repository.MongoOrdersRepository) gin.HandlerFunc {
-	return func(context *gin.Context) {
-		statusCode, response := repository.GetOrderById(context)
-		context.IndentedJSON(statusCode, response)
-	}
-}
