--- src/handlers/any_role/profile/update_profile.go
@@ -1,14 +1,0 @@
-package profile_handler
-
-import (
-	repository "comiditapp/api/database/repositories/users"
-
-	"github.com/gin-gonic/gin"
-)
-
-func UpdateProfile(repository *repository.MongoUsersRepository) gin.HandlerFunc {
-	return func(context *gin.Context) {
-		statusCode, response := repository.UpdateProfile(context)
-		context.IndentedJSON(statusCode, response)
-	}
-}
