--- src/handlers/any_role/restaurants/get_restaurant_by_id.go
@@ -1,14 +1,0 @@
-package restaurant_handler
-
-import (
-	repository "comiditapp/api/database/repositories/users"
-
-	"github.com/gin-gonic/gin"
-)
-
-func GetRestaurantById(repository *repository.MongoUsersRepository) gin.HandlerFunc {
-	return func(context *gin.Context) {
-		statusCode, response := repository.GetRestaurantById(context)
-		context.IndentedJSON(statusCode, response)
-	}
-}
