+++ src/handlers/auth/post_signin.go
@@ -1,0 +1,15 @@
+package handlers
+
+import (
+	repository "comiditapp/api/database/repositories/users"
+
+	"github.com/gin-gonic/gin"
+)
+
+func SignInUser(repository *repository.MongoUsersRepository) gin.HandlerFunc {
+	return func(context *gin.Context) {
+		statusCode, response := repository.SignInUser(context)
+		context.IndentedJSON(statusCode, response)
+
+	}
+}
