+++ src/handlers/auth/post_signout.go
@@ -1,0 +1,14 @@
+package handlers
+
+import (
+	repository "comiditapp/api/database/repositories/users"
+
+	"github.com/gin-gonic/gin"
+)
+
+func SignOutUser(repository *repository.MongoUsersRepository) gin.HandlerFunc {
+	return func(context *gin.Context) {
+		statusCode, response := repository.SignOutUser(context)
+		context.IndentedJSON(statusCode, response)
+	}
+}
