+++ src/handlers/auth/post_signup.go
@@ -1,0 +1,14 @@
+package handlers
+
+import (
+	repository "comiditapp/api/database/repositories/users"
+
+	"github.com/gin-gonic/gin"
+)
+
+func SignUpUser(repository *repository.MongoUsersRepository) gin.HandlerFunc {
+	return func(context *gin.Context) {
+		statusCode, response := repository.SignUpUser(context)
+		context.IndentedJSON(statusCode, response)
+	}
+}
