+++ src/handlers/home/get_home.go
@@ -1,0 +1,11 @@
+package handlers
+
+import (
+	"net/http"
+
+	"github.com/gin-gonic/gin"
+)
+
+func GetHome(c *gin.Context) {
+	c.String(http.StatusOK, "Welcome to comiditapp homepage!! 🍔")
+}
