+++ src/handlers/products/get_all_products.go
@@ -1,0 +1,14 @@
+package handlers
+
+import (
+	repository "comiditapp/api/database/repositories/users"
+
+	"github.com/gin-gonic/gin"
+)
+
+func FindProducts(repository *repository.MongoUsersRepository) gin.HandlerFunc {
+	return func(context *gin.Context) {
+		statusCode, response := repository.FindProducts(context)
+		context.IndentedJSON(statusCode, response)
+	}
+}
