+++ src/handlers/profile/delete_account.go
@@ -1,0 +1,14 @@
+package handlers
+
+import (
+	repository "comiditapp/api/database/repositories/users"
+
+	"github.com/gin-gonic/gin"
+)
+
+func DeleteAccount(repository *repository.MongoUsersRepository) gin.HandlerFunc {
+	return func(context *gin.Context) {
+		statusCode, response := repository.DeleteAccount(context)
+		context.IndentedJSON(statusCode, response)
+	}
+}
