--- src/handlers/restaurant_role/products/get_all_products.go
@@ -1,14 +1,0 @@
-package product_handler
-
-import (
-	repository "comiditapp/api/database/repositories/users"
-
-	"github.com/gin-gonic/gin"
-)
-
-func FindProducts(repository *repository.MongoUsersRepository) gin.HandlerFunc {
-	return func(context *gin.Context) {
-		statusCode, response := repository.FindProducts(context)
-		context.IndentedJSON(statusCode, response)
-	}
-}
