--- src/handlers/restaurant_role/products/update_product.go
@@ -1,14 +1,0 @@
-package product_handler
-
-import (
-	repository "comiditapp/api/database/repositories/users"
-
-	"github.com/gin-gonic/gin"
-)
-
-func UpdateProduct(repository *repository.MongoUsersRepository) gin.HandlerFunc {
-	return func(context *gin.Context) {
-		statusCode, response := repository.UpdateProduct(context)
-		context.IndentedJSON(statusCode, response)
-	}
-}
