+++ src/handlers/restaurants/get_all_restaurants.go
@@ -1,0 +1,14 @@
+package handlers
+
+import (
+	repository "comiditapp/api/database/repositories/users"
+
+	"github.com/gin-gonic/gin"
+)
+
+func FindRestaurants(repository *repository.MongoUsersRepository) gin.HandlerFunc {
+	return func(context *gin.Context) {
+		statusCode, response := repository.FindRestaurants(context)
+		context.IndentedJSON(statusCode, response)
+	}
+}
