--- src/middlewares/get_token_payload.go
@@ -1,8 +1,0 @@
-package middlewares
-
-import "github.com/dgrijalva/jwt-go"
-
-func GetTokenPayload(token string) jwt.Claims {
-	t, _ := jwt.Parse(token, nil)
-	return t.Claims.(jwt.MapClaims)
-}
